workers/python_worker/requirements.txt
pika
psycopg2-binary
requests
python-dotenv
prometheus_client
boto3
weasyprint
matplotlib
google-generativeai
openai
beautifulsoup4
lxml
tenacity
perplexity